backend/requirements.txt
asgiref==3.9.1
attrs==25.3.0
certifi==2025.10.5
charset-normalizer==3.4.4
Django==5.1.1
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
et_xmlfile==2.0.0
idna==3.11
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
openpyxl==3.1.5
pillow==11.3.0
psycopg2-binary==2.9.9
PyJWT==2.10.1
PyYAML==6.0.2
referencing==0.36.2
reportlab==4.4.4
requests==2.32.5
rpds-py==0.27.1
sqlparse==0.5.3
stripe==13.2.0
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.7.0
